Row Scan partition pruning test included in the existing to avoid extra time in dunit (precheckin) run
